# openapi.yaml
openapi: 3.0.3
info:
  title: Oblixa route contract
  version: 0.1.0
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
    CronSecret:
      type: http
      scheme: bearer
    WebhookSignature:
      type: apiKey
      in: header
      name: stripe-signature
paths:
...
  /api/integrations/oauth/callback:
    get:
      operationId: get_integrations_oauth_callback
      summary: GET /api/integrations/oauth/callback
      description: Generated route contract for src/app/api/integrations/oauth/callback/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...
  /api/report-packs/{id}/runs:
    get:
      operationId: get_report_packs_by_id_runs
      summary: GET /api/report-packs/{id}/runs
      description: Generated route contract for src/app/api/report-packs/[id]/runs/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...