# src/app/api/report-packs/[id]/runs/route.ts
import { NextResponse } from "next/server";
import { jsonNotFound, jsonProblem, jsonRateLimited, jsonUnauthorized } from "@/lib/http/problem";
import { getApiAuthContext } from "@/lib/contract-operations/api-auth";
import { requireApiWorkspaceEligibility } from "@/lib/product-surface/api-workspace-guard";
import { parseWorkspaceMode } from "@/lib/product-surface/context";
import { workspaceModeAllowsReportType } from "@/lib/product-surface/feature-registry";
import { getOrgSettingsJson } from "@/lib/assurance/org-settings";
import { escapeCsvCellForSpreadsheet } from "@/lib/csv-formula-safe";
import {
  contentDispositionAttachment,
  contentDispositionInline,
  sanitizeExportFileName,
  sanitizeExportFileNameToken,
} from "@/lib/security/export-filename";
import { recordApiRouteAuditEvent } from "@/lib/security/api-mutation-audit";
import { rejectUnsafeRouteParams } from "@/lib/security/route-params";
import { parseFixedEnumParam } from "@/lib/security/validation";
import { RATE_LIMITS, getClientIpFromRequest, rateLimitCheck } from "@/lib/rate-limit";

const ROUTE = "/api/report-packs/[id]/runs";

export async function GET(
  request: Request,
  { params }: { params: Promise<{ id: string }> }
) {
  const ctx = await getApiAuthContext();
  if (!ctx) return jsonUnauthorized(ROUTE);
  const modeGate = await requireApiWorkspaceEligibility({
    admin: ctx.admin,
    orgId: ctx.orgId,
    role: ctx.role,
    apiPath: "/api/report-packs/[id]/runs",
  });
  if (modeGate) return modeGate;

  void recordApiRouteAuditEvent(ctx.admin, {
    organizationId: ctx.orgId,
    actorUserId: ctx.userId,
    route: ROUTE,
    method: "GET",
    action: "api.sensitive_read_authorized",
  }).catch(() => undefined);

  const url = new URL(request.url);
  const format = parseFixedEnumParam(url.searchParams.get("format"), ["json", "csv", "html", "pdf"] as const, "json");
  const runId = url.searchParams.get("runId");

  const { id } = await params;

  const routeParamRejection = rejectUnsafeRouteParams({ id }, ["id"], "/api/report-packs/[id]/runs");

  if (routeParamRejection) return routeParamRejection;
  const ip = getClientIpFromRequest(request);
  const rl = await rateLimitCheck(`report-pack-runs:${ctx.userId}:${ip}:${id}`, RATE_LIMITS.reportPackRunsRead);
  if (!rl.ok) return jsonRateLimited(rl.retryAfterMs, ROUTE);

  const { data: pack } = await ctx.admin
    .from("report_packs")
    .select("id, name, report_type, annotations_json")
    .eq("id", id)
    .eq("organization_id", ctx.orgId)
    .maybeSingle();
  if (!pack) return jsonNotFound(ROUTE);

  const v6 = await getOrgSettingsJson(ctx.admin, ctx.orgId);
  const mode = parseWorkspaceMode(v6);
  if (!workspaceModeAllowsReportType(mode, String(pack.report_type ?? ""))) {
    return jsonNotFound(ROUTE);
  }

  const { data, error } = await ctx.admin
    .from("report_pack_runs")
    .select("id, status, started_at, completed_at, metrics_json, output_refs_json, error, created_at")
    .eq("organization_id", ctx.orgId)
    .eq("report_pack_id", id)
    .order("created_at", { ascending: false })
    .limit(100);
  if (error) {
    return jsonProblem(400, {
      error: error.message,
      code: "report_pack_runs_list_failed",
      diagnostic_id: "report_pack_runs_list_failed",
      route: ROUTE,
    });
  }

  const runs = data ?? [];
  const target = runId ? runs.find((r) => r.id === runId) : runs[0];
  const safePackId = sanitizeExportFileNameToken(id);

  if (format === "csv") {
    if (!target) return new NextResponse("No runs to export", { status: 404 });
    const metrics = (target.metrics_json ?? {}) as Record<string, unknown>;
    const csvFileName = sanitizeExportFileName(`report-pack-${safePackId}-run.csv`);
    const lines = ["key,value"];
    for (const [k, v] of Object.entries(metrics)) {
      lines.push(
        `${escapeCsvCellForSpreadsheet(k)},${escapeCsvCellForSpreadsheet(
          typeof v === "object" ? JSON.stringify(v) : String(v)
        )}`
      );
    }
    return new NextResponse(lines.join("\r\n"), {
      status: 200,
      headers: {
        "content-type": "text/csv; charset=utf-8",
        "content-disposition": contentDispositionAttachment(csvFileName),
        "cache-control": "private, no-store",
      },
    });
  }

  if (format === "html" || format === "pdf") {
    if (!target) return new NextResponse("No runs to export", { status: 404 });
    const metrics = (target.metrics_json ?? {}) as Record<string, unknown>;
    const annotations = (pack.annotations_json as unknown[] | null) ?? [];
    const rows = Object.entries(metrics)
      .map(
        ([k, v]) =>
          `<tr><th style="text-align:left;padding:6px;border-bottom:1px solid #e4e4e7">${escapeHtml(k)}</th><td style="padding:6px;border-bottom:1px solid #e4e4e7">${escapeHtml(
            typeof v === "object" ? JSON.stringify(v) : String(v)
          )}</td></tr>`
      )
      .join("");
    const annRows = annotations
      .map((a, i) => `<li>${escapeHtml(typeof a === "object" ? JSON.stringify(a) : String(a))} (${i + 1})</li>`)
      .join("");
    const htmlFileName = sanitizeExportFileName(`report-pack-${safePackId}.html`);
    const html = `<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"/><title>Report pack — ${escapeHtml(
      pack.name ?? id
    )}</title><style>body{font-family:system-ui,sans-serif;padding:24px;color:#18181b}h1{font-size:20px}table{border-collapse:collapse;width:100%;max-width:720px}@media print{body{padding:12px}}</style></head><body><h1>${escapeHtml(
      pack.name ?? "Report pack"
    )}</h1><p style="color:#52525b">Type: ${escapeHtml(pack.report_type)} · Run: ${escapeHtml(
      target.id
    )} · Generated ${escapeHtml(target.completed_at ?? target.created_at)}</p><table><tbody>${rows}</tbody></table>${
      annRows
        ? `<h2 style="margin-top:24px;font-size:16px">Annotations</h2><ol>${annRows}</ol>`
        : ""
    }<p style="margin-top:32px;font-size:12px;color:#71717a">Print this page or save as PDF from your browser.</p></body></html>`;
    const headers: Record<string, string> = {
      "content-type": "text/html; charset=utf-8",
      "content-disposition": contentDispositionInline(htmlFileName),
      "cache-control": "private, no-store",
    };
    if (format === "pdf") {
      headers["x-oblixa-print-hint"] = "Save as PDF from browser print dialog";
    }
    return new NextResponse(html, { status: 200, headers });
  }

  return NextResponse.json({ runs }, { headers: { "Cache-Control": "private, no-store" } });
}

function escapeHtml(s: string) {
  return s
    .replace(/&/g, "&amp;")
    .replace(/</g, "&lt;")
    .replace(/>/g, "&gt;")
    .replace(/"/g, "&quot;");
}